Removed unused import Refactored out the button to be reused elsewhere
